refactor: 필드 타입 리팩토링
- not null 필드는 primitive 타입으로 수정
- nullable 또는 id(pk) 필드는 Wrapper 타입으로 수정

+++ mobidic-domain/src/main/java/com/kthowns/mobidic/domain/word/model/WordDetail.java
@@ -11,8 +11,8 @@ import java.util.UUID;
 public record WordDetail(
         UUID id,
         String expression,
-        double difficulty,
-        double accuracy,
+        long difficulty,
+        long accuracy,
         boolean isLearned,
         List<Definition> definitions,
         LocalDateTime createdAt
